==> stanford_dog/00 - sample_data.ipynb <==
from fastai.vision.all import *
from scipy.io import loadmat

datapath = Path("stanford-dogs-dataset/")

list(datapath.ls())

test_filenames = [o[0][0] for o in loadmat(datapath/'test_list.mat')['file_list']]
train_filenames = [o[0][0] for o in loadmat(datapath/'train_list.mat')['file_list']]
test_filenames = [o[0][0] for o in loadmat(datapath/'test_list.mat')['file_list']]

len(train_filenames), len(test_filenames)

train_labels = L(train_filenames).map(lambda o: o.split("/")[0])

train_df = pd.DataFrame({"filenames":train_filenames, "labels": train_labels})

test_labels = L(test_filenames).map(lambda o: o.split("/")[0])

test_df = pd.DataFrame({"filenames":test_filenames, "labels": test_labels})

train_df.to_csv(datapath/"train.csv", index=False)
test_df.to_csv(datapath/"test.csv", index=False)

train_df.shape, test_df.shape

### Subsample Train (40 train, 10 valid per class) - 4800/1200

class2filename = dict(train_df.groupby("labels")['filenames'].apply(lambda o: np.random.choice(o, size=50, replace=False)))

sample_train_df = []
for c in class2filename:
    for v in class2filename[c]:
        sample_train_df.append((c,v))

sample_train_df = pd.DataFrame(sample_train_df, columns=['label', 'filename'])

valid_filenames = sample_train_df.groupby("label").head(10)['filename']
sample_train_df['split'] = 'train'
sample_train_df.loc[sample_train_df.filename.isin(valid_filenames), 'split'] = 'valid'

sample_train_df.to_csv(datapath/"sample_train.csv", index=False)
